requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.6
coreschema==0.0.4
Django==2.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-guardian==1.4.9
djangorestframework==3.7.3
httpie==0.9.9
httplib2==0.10.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.10
MarkupSafe==1.0
Pygments==2.2.0
pytz==2017.3
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
